+++ R/approx.R
@@ -45,6 +45,37 @@ mvqt_pois <- function(L, k, mean, Q) {
 }
 
 
+mvgamma_pois_approx <- function(k, mean, Q) {
+  invtau <- 1.0/diag(Q)
+  m <- exp(mean + 0.5*invtau)
+  mm <- m*m
+  v <- (exp(invtau) - 1.0) * mm
+  alpha <- mm / v + replace(k, is.na(k), 0)
+  beta <- m / v + (!is.na(k))
+  gu_params(alpha = alpha, beta = beta)
+}
+
+mvgamma_pois <- function(L, mult, k, mean, Q) {
+
+  pois_LL_mv <- function(L, k, mean, Q) {
+    sum((k * L - exp(L))[!is.na(k)]) - 0.5*as.vector(t(L - mean) %*% Q %*% (L - mean))
+  }
+
+  tmp <- mvgamma_pois_approx(k, mean, Q)
+  alpha2 <- tmp$alpha / mult
+  scale2 <- mult / tmp$beta
+
+  proposal <- rgamma(length(L), alpha2, scale = scale2)
+  lproposal <- log(proposal)
+
+  ratio <- pois_LL_mv(lproposal, k, mean, Q)
+  ratio <- ratio - sum((alpha2 - 1.0)*lproposal - proposal/scale2)
+  ratio <- ratio - pois_LL_mv(L, k, mean, Q)
+  ratio <- ratio + sum((alpha2 - 1.0)*L - exp(L)/scale2)
+  accept <- accept_reject(ratio)
+  if(accept) c(1, lproposal) else c(0, L)
+}
+
 
 #' @rdname approximations
 #' @export