+++ R/pois.R
@@ -9,9 +9,12 @@
 #' @param ncontract The maximum number of contractions. If this is exceeded, the original value is returned
 #' @param method The method to use to propose new values. \code{"normal"} proposes normal deviates from the current
 #'   value. \code{"uniform"} proposes uniform deviates. \code{"gamma"} does moment-matching on the mean and variance
-#'   to propose a conjugate gamma proposal. \code{"quadratic taylor"} proposes using a second-order Taylor
+#'   to propose a conjugate gamma proposal. \code{"mv gamma"} does the same, but ignores off-diagonal elements of the precision matrix
+#'   in the proposal (for speed's sake; this may result in low acceptance rates when elements are highly correlated).
+#'    \code{"quadratic taylor"} proposes using a second-order Taylor
 #'   approximation of the log-density at the current value, which amounts to a normal proposal with mean (usually) not equal
-#'   to the current value.
+#'   to the current value. \code{"mv quadratic taylor"} does the same, but uses a multivariate normal approximation.
+#'   Both \code{"mv gamma"} and \code{"mv quadratic taylor"} accept or reject an entire row at a time.
 #' @param width For \code{"normal"} proposals, the standard deviation(s) of proposals. For \code{"uniform"} proposals,
 #' the half-width of the uniform proposal interval. For \code{"slice"}, the width of each expansion (to the right and left each).
 #' For \code{"gamma"} a scaling factor to increase the variance of the proposal.
@@ -32,7 +35,7 @@
 #' @seealso \url{https://en.wikipedia.org/wiki/Slice_sampling},
 #' \url{https://en.wikipedia.org/wiki/Metropolis–Hastings_algorithm}, \url{https://arxiv.org/pdf/1308.0657.pdf}
 #' @export
-sample_pois_reg <- function(L, k, mean, precision, method = c("slice", "normal", "uniform", "gamma", "quadratic taylor", "mv quadratic taylor"), ...,
+sample_pois_reg <- function(L, k, mean, precision, method = c("slice", "normal", "uniform", "gamma", "mv gamma", "quadratic taylor", "mv quadratic taylor"), ...,
                             width = 1, nexpand = 10, ncontract = 100) {
   method <- match.arg(method)
 
@@ -67,14 +70,19 @@ sample_pois_reg <- function(L, k, mean, precision, method = c("slice", "normal",
       out <- slice_sample_pois(L = L, k = k, k_na = is.na(k), mean = mean,
                                precision = precision, w = width, nexpand = nexpand, ncontract = ncontract)
     }
-  } else if(method %in% c("normal", "uniform", "gamma")) {
-    width <- check_one_or_all(width, length(L))
+  } else if(method %in% c("normal", "uniform", "gamma", "quadratic taylor")) {
     if(method %in% c("normal", "uniform")) {
+      width <- check_one_or_all(width, length(L))
       prop <- L + if(method == "normal") stats::rnorm(length(L), 0, width) else stats::runif(length(L), -width, width)
       m <- 0L
-    } else {
+    } else if(method == "gamma") {
+      width <- check_one_or_all(width, length(L))
       prop <- width
       m <- 2L
+    } else if(method == "quadratic taylor") {
+      if(!missing(width)) warning("'width' is being ignored for this method.")
+      prop <- L # this is ignored
+      m <- 1L
     }
 
     if(is.matrix(precision)) {
@@ -84,30 +92,41 @@ sample_pois_reg <- function(L, k, mean, precision, method = c("slice", "normal",
     } else {
       out <- mh_pois(method = m, L = L, proposal = prop, k = k, k_na = is.na(k), mean = mean, precision = precision)
     }
-  } else if(method %in% c("quadratic taylor", "mv quadratic taylor")) {
+  } else if(method == "mv gamma") {
+    if(is.matrix(precision)) {
+      width <- check_one_or_all(width, length(L))
+      dim(width) <- dim(L)
+      use_norm[use_norm] <- seq_len(sum(use_norm))
+      tmp <- t(vapply(seq_len(nrow(L)), function(i) {
+        if(use_norm[i] > 0) {
+          return(c(1, norm[use_norm[i], ]))
+        }
+        mvgamma_pois(L[i, ], width[i, ], k[i, ], mean[i, ], precision)
+      }, numeric(ncol(L)+1)))
+      out <- tmp[, -1]
+      attr(out, "accept") <- array(as.logical(tmp[, 1]), dim = dim(L))
+    } else {
+      warning("'mv gamma' is being interpreted as 'gamma' because 'precision' is not a matrix.")
+      out <- mh_pois(method = 2L, L = L, proposal = width, k = k, k_na = is.na(k), mean = mean, precision = precision)
+    }
+  } else if(method == "mv quadratic taylor") {
     if(!missing(width)) warning("'width' is being ignored for this method.")
 
     if(is.matrix(precision)) {
-      if(method == "mv quadratic taylor") {
-        use_norm[use_norm] <- seq_len(sum(use_norm))
-        tmp <- t(vapply(seq_len(nrow(L)), function(i) {
-          if(use_norm[i] > 0) {
-            return(c(1, norm[use_norm[i], ]))
-          }
-          mvqt_pois(L[i, ], k[i, ], mean[i, ], precision)
-        }, numeric(ncol(L)+1)))
-        out <- tmp[, -1]
-        attr(out, "accept") <- array(as.logical(tmp[, 1]), dim = dim(L))
-
-      } else {
-        out <- mh_pois_mv(method = 1L, L = L, proposal = matrix(NA_real_), k = k, k_na = is.na(k), mean = mean,
-                          Q = precision, use_norm = use_norm, norm = norm)
-      }
+      use_norm[use_norm] <- seq_len(sum(use_norm))
+      tmp <- t(vapply(seq_len(nrow(L)), function(i) {
+        if(use_norm[i] > 0) {
+          return(c(1, norm[use_norm[i], ]))
+        }
+        mvqt_pois(L[i, ], k[i, ], mean[i, ], precision)
+      }, numeric(ncol(L)+1)))
+      out <- tmp[, -1]
+      attr(out, "accept") <- array(as.logical(tmp[, 1]), dim = dim(L))
     } else {
       if(method == "mv quadratic taylor") {
         warning("'mv quadratic taylor' is being interpreted as 'quadratic taylor' because 'precision' is not a matrix.")
       }
-      out <- mh_pois(method = 1L, L = L, proposal = NA_real_, k = k, k_na = is.na(k), mean = mean, precision = precision)
+      out <- mh_pois(method = 1L, L = L, proposal = L, k = k, k_na = is.na(k), mean = mean, precision = precision)
     }
   }